requirements/base.txt
Django==3.1
Pillow==7.2.0

djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
pyjwt==2.3.0

django-extensions==3.0.4
django-constance[database]==2.7.0
django-debug-toolbar==3.2.4
 psycopg2-binary==2.9.3


